# Performance: Implement I/O and protocol-level optimizations for HTTP/2 client